// frontend/src/__generated__/models/Body_add_smart_collection_api_collections_smart_post.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Body_add_smart_collection_api_collections_smart_post = {
    name?: string;
    description?: string;
    /**
     * Smart collection filters as a JSON string.
     */
    filter_criteria?: string;
};