// frontend/src/__generated__/models/Body_update_smart_collection_api_collections_smart__id__put.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Body_update_smart_collection_api_collections_smart__id__put = {
    name?: (string | null);
    description?: (string | null);
    /**
     * Updated smart collection filters as a JSON string.
     */
    filter_criteria?: (string | null);
};